+++ example/lib/annotation_order_maps.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart'; // ignore: unnecessary_import
 import 'package:vietmap_flutter_gl/vietmap_flutter_gl.dart';
 
+import 'constant.dart';
 import 'page.dart';
 import 'util.dart';
 
@@ -57,6 +58,7 @@ class _AnnotationOrderBodyState extends State<AnnotationOrderBody> {
                         AnnotationType.circle,
                         AnnotationType.fill,
                       ],
+                      styleString: YOUR_STYLE_URL_HERE,
                     ),
                   ),
                 ),
@@ -79,6 +81,7 @@ class _AnnotationOrderBodyState extends State<AnnotationOrderBody> {
                     width: 250.0,
                     height: 250.0,
                     child: VietmapGL(
+                      styleString: YOUR_STYLE_URL_HERE,
                       onMapCreated: onMapCreatedTwo,
                       onStyleLoadedCallback: () => onStyleLoaded(controllerTwo),
                       initialCameraPosition: CameraPosition(