+++ example/lib/click_annotations.dart
@@ -5,6 +5,7 @@
 import 'package:flutter/material.dart';
 import 'package:vietmap_flutter_gl/vietmap_flutter_gl.dart';
 
+import 'constant.dart';
 import 'page.dart';
 import 'util.dart';
 
@@ -134,6 +135,7 @@ class ClickAnnotationBodyState extends State<ClickAnnotationBody> {
   @override
   Widget build(BuildContext context) {
     return VietmapGL(
+              styleString: YOUR_STYLE_URL_HERE,
       annotationOrder: [
         AnnotationType.fill,
         AnnotationType.line,