+++ example/lib/constant.dart
@@ -1,0 +1,2 @@
+const String YOUR_STYLE_URL_HERE =
+    "https://run.mocky.io/v3/961aaa3a-f380-46be-9159-09cc985d9326";